get_css now takes any Sublime scope and returns CSS
get_css no longer requires parameter to be as a CSS class and can take
any Sublime scope regardless of complexity.  Also can specify if you
want a valid background color to be returned always, or only when
explicitly set due to scopes. Removed a lot of related code that is no
longer needed with this change.